refactor: migrate blockSettings:tableSelector
