+++ packages/core/client/src/modules/blocks/data-blocks/table-selector/blockSettings:tableSelector.ts
@@ -3,11 +3,15 @@ import { ISchema, useField, useFieldSchema } from '@formily/react';
 import { cloneDeep } from 'lodash';
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
-import { SchemaSettings } from '../../../application/schema-settings/SchemaSettings';
-import { recursiveParent, useFormBlockContext, useTableSelectorContext } from '../../../block-provider';
-import { useCollection_deprecated, useCollectionManager_deprecated, useSortFields } from '../../../collection-manager';
-import { removeNullCondition, useDesignable } from '../../../schema-component';
-import { SchemaSettingsDataScope } from '../../../schema-settings/SchemaSettingsDataScope';
+import { SchemaSettings } from '../../../../application/schema-settings/SchemaSettings';
+import { recursiveParent, useFormBlockContext, useTableSelectorContext } from '../../../../block-provider';
+import {
+  useCollection_deprecated,
+  useCollectionManager_deprecated,
+  useSortFields,
+} from '../../../../collection-manager';
+import { removeNullCondition, useDesignable } from '../../../../schema-component';
+import { SchemaSettingsDataScope } from '../../../../schema-settings/SchemaSettingsDataScope';
 
 export const dataSelectorBlockSettings = new SchemaSettings({
   name: 'blockSettings:tableSelector',
